Imp: wrap indent takes same characters into consideration as auto indent
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk@496 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/CEditView_Command.cpp
@@ -184,7 +184,7 @@ BOOL CEditView::HandleCommand(
 			else {
 				//	一文字前にSJISの1バイト目が来ている
 				if( _IS_SJIS_2( (unsigned char)lparam1 )){
-					Command_IME_CHAR( ucSjis1 << 8 | lparam1 );
+					Command_IME_CHAR( (WORD)(ucSjis1 << 8 | lparam1 ));
 					ucSjis1 = 0;
 					break;
 				}
